+++ receipt-service/src/main/java/com/budgetmate/receipt/ReceiptServiceApplication.java
@@ -1,0 +1,13 @@
+package com.budgetmate.receipt;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReceiptServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReceiptServiceApplication.class, args);
+	}
+
+}
