+++ receipt-service/src/main/java/com/budgetmate/receipt/dto/OcrRequestDto.java
@@ -1,0 +1,11 @@
+package com.budgetmate.receipt.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.web.multipart.MultipartFile;
+
+@Getter
+@Setter
+public class OcrRequestDto {
+    private MultipartFile image;
+}
