+++ receipt-service/src/main/java/com/budgetmate/receipt/repository/ReceiptRepository.java
@@ -1,0 +1,9 @@
+package com.budgetmate.receipt.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.budgetmate.receipt.entity.ReceiptEntity;
+
+public interface ReceiptRepository extends JpaRepository <ReceiptEntity, Long>{
+
+}
