--- receipt-service/src/main/java/me/khs/ReceiptServiceApplication.java
@@ -1,13 +1,0 @@
-package me.khs;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReceiptServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ReceiptServiceApplication.class, args);
-	}
-
-}
