--- receipt-service/src/main/java/me/khs/repository/ReceiptRepository.java
@@ -1,9 +1,0 @@
-package me.khs.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import me.khs.entity.ReceiptEntity;
-
-public interface ReceiptRepository extends JpaRepository <ReceiptEntity, Long>{
-
-}
