+++ receipt-service/src/main/resources/application.yml
@@ -1,35 +1,12 @@
 #도커 연동 전.
-server:
-  port: 8082
 spring:
-  datasource:
-    url: jdbc:mysql://localhost:3306/budget_db?allowPublicKeyRetrieval=true&useSSL=false&serverTimezone=Asia/Seoul&allowMultiQueries=true
-    username: budget_user
-    password: 1234
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
- # config 설정을 끄기.
-  cloud:
-    config:
-      enabled: false
-# gateway 설정을 끄기.
-    gateway:
-      enabled: false
-# eureka 설정
-    discovery:
-      enabled: true
-  application:
-    name: receipt-service
-# circuit-breaker 설정 끄기.
-    circuitbreaker:
-      enabled: false
-  sql:
-    init:
-      mode: always
-  logging:
-    level:
-      root: debug
-      org.springframework: debug
+  config:
+    import: optional:configserver:http://localhost:8888
+
+  profiles:
+    active: ${SPRING_PROFILES_ACTIVE:local}
+
+eureka:
+  client:
+    register-with-eureka: true
+    fetch-registry: true
