+++ receipt-service/src/test/java/com/budgetmate/receipt/ReceiptServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.budgetmate.receipt;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ReceiptServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
