--- receipt-service/src/test/java/me/khs/ReceiptServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package me.khs;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReceiptServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
